requirements.txt
fastapi
pandas
requests
playwright
uvicorn
python-dotenv
